cleaning up code, moved main.dart and tabbed appbar code into separate files for ease of reading
