+++ crypto_balance/lib/balance.dart
@@ -1,4 +1,5 @@
 import 'package:crypto_balance/tabbedAppbar.dart';
+import 'package:crypto_balance/assets/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:sqflite/sqflite.dart';
