+++ crypto_balance/lib/tabbedAppbar.dart
@@ -4,94 +4,16 @@
 //import 'package:crypto_balance/Jerid.dart';
 import 'package:crypto_balance/main.dart';
 import 'package:flutter/material.dart';
-import 'package:crypto_balance/references.dart';
-import 'package:crypto_balance/balance.dart';
 //below: for global database
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart' as p;
 import 'package:flutter/widgets.dart';
 
-
-
-String currencySelection = currUSD;
+import 'package:crypto_balance/pricesTab.dart';
+import 'package:crypto_balance/balance.dart';
+import 'package:crypto_balance/references.dart';
 
 void main()=> runApp(MyApp1());
-
-//set up string names
-const String currUSD = "USD";
-const String currJPY = "JPY";
-const String currGBP = "GBP";
-const String currMXN = "MXN";
-const String currZAR = "ZAR";
-const String currCNY = "CNY";
-const String currEUR = "EUR";
-const String currKRW = "KRW";
-const String currINR = "INR";
-const String currTHB = "THB";
-const String currPHP = "PHP";
-
-//important global variables accessible across tabs
-Map globalCryptoPrice = Map<String, dynamic>();
-Map globalConvFac = Map<String, dynamic>();
-String globalCurr = currUSD;
-final String dbPath = "balance_db.db";
-
-class Currency{
-  final Text current;
-  final String acronym;
-  const Currency({
-    this.current,
-    this.acronym,
-  });
-}
-
-const List<Currency> Currencies = <Currency>[
-  const Currency(
-    current: const Text('USD',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currUSD,
-  ),
-  const Currency(
-    current: const Text('YEN',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currJPY,
-  ),
-  const Currency(
-    current: const Text('Pound',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currGBP,
-  ),
-  const Currency(
-    current: const Text('Peso',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currMXN,
-  ),
-  const Currency(
-    current: const Text('Rand',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currZAR,
-  ),
-  const Currency(
-    current: const Text('Rupee',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currINR,
-  ),
-  const Currency(
-    current: const Text('Yaun',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currCNY,
-  ),
-  const Currency(
-    current: const Text('Euro',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currEUR,
-  ),
-  const Currency(
-    current: const Text('Won',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currKRW,
-  ),
-  const Currency(
-    current: const Text('Baht',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currTHB,
-  ),
-  const Currency(
-    current: const Text('Philippine',style: TextStyle(color: Color(0xff3E0CA9)),),
-    acronym: currPHP,
-  ),
-
-];
 
 //ToDo: add author to references
 /*
@@ -127,8 +49,12 @@ class _TabContainerBottomState extends State<TabContainerBottom> {
   void initState() {
     super.initState();
     screenList = [
+      //todo: change names to meaningful
+      //MyApp: prices list
       MyApp(),
+      //MyApp4: balances
       MyApp4(),
+      //MyApp2: references tab
       MyApp2(),
     ];
   }
@@ -196,69 +122,5 @@ class _TabContainerBottomState extends State<TabContainerBottom> {
   }
 }
 
-
-//ToDo: delete below? doesn't matter if reload? or implement
-//below: works, and can be used to not reload every tab every time, but need to implement everything in medium article after index list to preven
-//all tabs running at startup and causing permission issue
-//class TabContainerIndexedStack extends StatefulWidget {
-//  TabContainerIndexedStack({Key key}) : super(key: key);
-//
-//  @override
-//  _TabContainerIndexedStackState createState() => _TabContainerIndexedStackState();
-//}
-//
-
-//class TabContainerIndexedStack extends StatefulWidget {
-//  TabContainerIndexedStack({Key key}) : super(key: key);
-//
-//  @override
-//  _TabContainerIndexedStackState createState() => _TabContainerIndexedStackState();
-//}
-//
-//class _TabContainerIndexedStackState extends State<TabContainerIndexedStack> {
-//  int tabIndex = 0;
-//  List<Widget> screenList;
-//  @override
-//  void initState() {
-//    super.initState();
-//    screenList = [
-//    MyApp(),
-//    MyApp4(),
-//    MyApp2(),
-//    ];
-//  }
-//
-//  @override
-//  Widget build(BuildContext context) {
-//    return MaterialApp(
-//      home: Scaffold(
-//        body: IndexedStack(index: tabIndex, children: screenList),
-//        bottomNavigationBar: BottomNavigationBar(
-//          currentIndex: tabIndex,
-//          onTap: (int index) {
-//            setState((){
-//              tabIndex = index;
-//            });
-//          },
-//          items: [
-//              BottomNavigationBarItem(
-//                icon: Icon(Icons.attach_money),
-//                title: Text('Prices'),
-//              ),
-//              BottomNavigationBarItem(
-//                icon: Icon(Icons.timeline),
-//                title: Text('Balance'),
-//              ),
-//              BottomNavigationBarItem(
-//                icon: Icon(Icons.help),
-//                title: Text('About'),
-//              ),
-//              ],
-//        ),
-//        backgroundColor: Colors.blueGrey,
-//      ),
-//    );
-//  }
-//}
